# swagger.json
{"openapi":"3.0.0","paths":{"/notifications/messages":{"get":{"operationId":"NotificationsController_getNotifications","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseGeneric"}}}}},"tags":["Notificatios"],"security":[{"bearer":[]}]}},"/notifications/add-notifications":{"post":{"operationId":"NotificationsController_addNotifications","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseGeneric"}}}}},"tags":["Notificatios"],"security":[{"bearer":[]}]}},"/auth/register":{"post":{"operationId":"AuthController_registerUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseGeneric"}}}}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseGeneric"}}}}},"tags":["Auth"]}}},"info":{"title":"Jgaad API","description":"api documentations","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ResponseGeneric":{"type":"object","properties":{"succeeded":{"type":"boolean"},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"data":{"type":"object"}},"required":["succeeded","message","errors","data"]},"NotificationDto":{"type":"object","properties":{"email":{"type":"string"},"notifications":{"type":"array","items":{"type":"string"}}},"required":["email","notifications"]},"UserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]}}}}